Added BaseMXTouchContainer Pulled common code from MXTouchContainer and MXTouchSlideoutContainer into it
