+++ MonoCross.Touch/MXTouchContainer.cs
@@ -117,13 +117,10 @@ namespace MonoCross.Touch
 		//public String Icon { get; set; }
 	}
 	
-	public class MXTouchContainer: MXContainer
+	public class MXTouchContainer: BaseMXTouchContainer
 	{
-		MXTouchNavigation _touchNavigation;
-		UIWindow _window;
-		UIApplicationDelegate _appDelegate;
-		LoadingView _loadingView;
-		SplashViewController _splashViewController = null;
+		MXTouchNavigation touchNavigation;
+		LoadingView loadingView;
 
 		public static RenderLayerDelegate RenderLayer { get; set; }
 
@@ -131,36 +128,16 @@ namespace MonoCross.Touch
 
 
 
-		private MXTouchContainer (MXApplication theApp, UIApplicationDelegate appDelegate, UIWindow window): base(theApp)
+		private MXTouchContainer (MXApplication theApp, UIApplicationDelegate appDelegate, UIWindow window): base(theApp, appDelegate, window)
 		{
-			_appDelegate = appDelegate;
-			_touchNavigation = new MXTouchNavigation(_appDelegate, window);
-			_window = window;
-			
+			touchNavigation = new MXTouchNavigation(appDelegate, window);
+
 			ViewGroups = new List<MXTouchViewGroup>();
 		}
 
 		public List<MXTouchViewGroup> ViewGroups { get; private set; }
 
-		private void StartApplication()
-		{
-			if (_window.Subviews.Length == 0)
-			{
-				// toss in a temporary view until async initialization is complete
-				string bitmapFile = string.Empty;
-				MXTouchContainerOptions options = Attribute.GetCustomAttribute(_appDelegate.GetType(), typeof(MXTouchContainerOptions)) as MXTouchContainerOptions;
-				if (options != null) {
-					bitmapFile = options.SplashBitmap;
-				}
 
-				if (!String.IsNullOrEmpty(bitmapFile))
-				{
-					_splashViewController = new SplashViewController(bitmapFile);
-					_window.AddSubview(_splashViewController.View);
-					_window.MakeKeyAndVisible();
-				}
-			}
-		}
 		
 		public static void Initialize(MXApplication theApp, UIApplicationDelegate appDelegate, UIWindow window)
 		{
@@ -171,8 +148,8 @@ namespace MonoCross.Touch
 			thisContainer.StartApplication();
 		}
 		
-		public UINavigationController MasterNavigationController { get { return _touchNavigation.MasterNavigationController; } }
-		public UINavigationController DetailNavigationController { get { return _touchNavigation.DetailNavigationController; } }
+		public UINavigationController MasterNavigationController { get { return touchNavigation.MasterNavigationController; } }
+		public UINavigationController DetailNavigationController { get { return touchNavigation.DetailNavigationController; } }
 		
 		public override void Redirect(string url)
 		{
@@ -197,7 +174,7 @@ namespace MonoCross.Touch
 		{
 			Console.WriteLine("Controller Load End");
 			
-			_appDelegate.InvokeOnMainThread( delegate {
+			appDelegate.InvokeOnMainThread( delegate {
 				LoadViewForController(fromView, controller, viewPerspective);
 				if (ControllerLoadComplete != null)
 					ControllerLoadComplete(controller);
@@ -247,12 +224,12 @@ namespace MonoCross.Touch
 		{
 			if (_firstView)
 			{
-				foreach (var view in _window.Subviews)
+				foreach (var view in window.Subviews)
 					view.RemoveFromSuperview();
 				
 				_firstView = false;
-				_window.Add(_touchNavigation.View);
-				_window.MakeKeyAndVisible();
+				window.Add(touchNavigation.View);
+				window.MakeKeyAndVisible();
 			}
 		}
 		
@@ -297,7 +274,7 @@ namespace MonoCross.Touch
 			if (navigationContext == ViewNavigationContext.Modal)
 			{
 				// treat as a modal/popup view
-				_touchNavigation.PushToModel(viewController);
+				touchNavigation.PushToModel(viewController);
 			}
 			else if (navigationContext == ViewNavigationContext.InContext)
 			{
@@ -330,7 +307,7 @@ namespace MonoCross.Touch
 				if (viewGroup != null)
 				{
 					// activate the group!
-					_touchNavigation.PushToViewGroup(viewGroup, viewGroupItem, controller.View as UIViewController);
+					touchNavigation.PushToViewGroup(viewGroup, viewGroupItem, controller.View as UIViewController);
 
 				}
 				else
@@ -338,10 +315,10 @@ namespace MonoCross.Touch
 					switch (navigationContext)
 					{
 					case ViewNavigationContext.Detail:
-						_touchNavigation.PushToDetail(viewController);
+						touchNavigation.PushToDetail(viewController);
 						break;
 					case ViewNavigationContext.Master:
-						_touchNavigation.PushToMaster(viewController);
+						touchNavigation.PushToMaster(viewController);
 						break;
 					}
 				}
