+++ MonoCross.Touch/MonoCross.Touch.csproj
@@ -75,6 +75,7 @@
     <Compile Include="MXSlideoutContainer.cs" />
     <Compile Include="SlideoutNavigationController.cs" />
     <Compile Include="MXTouchSlideoutContainer.cs" />
+    <Compile Include="BaseMXTouchContainer.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MonoCross.Navigation\MonoCross.Navigation.MT.csproj">
